Separating concerns in the models folder
Created a main photo repo file, where all pictures would be saved and assigned a unique ID.
This ID would be used to reference the photo in sections of the website (e.g. pics for rooms, pics for shareda amenities, etc.).
Lastly, created files for each section of the website that would need photos.